EDX-1774 - Fixes next buttons for steps 3 and 4, and removes next functionality for now until step 6 is created.
