+++ frontend/src/components/sdcCollection/StepFiveSchoolContacts.vue
@@ -51,8 +51,6 @@ import SchoolContactsForm from '../common/forms/SchoolContactsForm.vue';
 import {isContactCurrent} from '../../utils/institute/status';
 import { mapState } from 'pinia';
 import { useSdcCollectionStore } from '../../store/modules/sdcCollection';
-import ApiService from '../../common/apiService';
-import { ApiRoutes } from '../../utils/constants';
   
 export default {
   name: 'StepFiveSchoolContacts',
@@ -86,23 +84,7 @@ export default {
     next() {
       if(this.currentStepInCollectionProcess.isComplete) {
         this.$emit('next');
-      } else {
-        this.markStepAsComplete();
-      }     
-    },
-    markStepAsComplete() {
-      let updateCollection = {
-        schoolCollection: this.schoolCollectionObject,
-        status: 'SCH_C_VRFD'
-      };
-      ApiService.apiAxios.put(ApiRoutes.sdc.BASE_URL + '/' + this.sdcSchoolCollectionID, updateCollection)
-        .then(() => {
-          this.$emit('next');
-        })
-        .catch(error => {
-          console.error(error);
-          this.setFailureAlert(error?.response?.data?.message ? error?.response?.data?.message : 'An error occurred while verifying school contact details. Please try again later.');
-        }); 
+      }
     },
     checkIfPrincipalContactExists(contacts) {
       let contact = contacts.filter(contact => contact.schoolContactTypeCode === 'PRINCIPAL' && isContactCurrent(contact));
