+++ frontend/src/components/sdcCollection/StepFourSchoolDetails.vue
@@ -97,7 +97,7 @@ export default {
     markStepAsComplete() {
       let updateCollection = {
         schoolCollection: this.schoolCollectionObject,
-        status: 'SCH_D_VRFD'
+        status: 'SCH_C_VRFD'
       };
       ApiService.apiAxios.put(ApiRoutes.sdc.BASE_URL + '/' + this.sdcSchoolCollectionID, updateCollection)
         .then(() => {
